# backend/cmd/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API built with Golang using Clean Code architecture. \\nGithub: [https://github.com/shion0625/backend].",
    "title": "E-commerce Application Backend API",
    "contact": {
      "name": "For API Support",
      "email": "nikhilnarayanan623@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
  "paths": {
...
    "/admin/orders/": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to change order status",
        "tags": ["Admin Orders"],
        "summary": "Change order status (Admin)",
        "operationId": "UpdateOrderStatus",
        "parameters": [
          {
            "description": "input field",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.UpdateOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully order status updated",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/orders/all": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all orders",
        "tags": ["Admin Orders"],
        "summary": "Get all orders (Admin)",
        "operationId": "GetAllShopOrders",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all shop orders",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "204": {
            "description": "No shop order found",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to find all shop orders",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/orders/returns": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all order returns",
        "tags": ["Admin Orders"],
        "summary": "Get all order returns (Admin)",
        "operationId": "GetAllOrderReturns",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count Of Order",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found all order returns",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to find all order returns",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/orders/returns/pending": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all pending returns",
        "tags": ["Admin Orders"],
        "summary": "Get all pending returns (Admin)",
        "operationId": "GetAllPendingReturns",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count Of Order",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found all pending orders return requests",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to find all pending order return requests",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to change status of return requested orders",
        "tags": ["Admin Orders"],
        "summary": "Change return request status (Admin)",
        "operationId": "UpdateReturnRequest",
        "parameters": [
          {
            "description": "Input Fields",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.UpdateOrderReturn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully order_response updated",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/orders/statuses": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all available order statuses",
        "tags": ["Admin Orders"],
        "summary": "Get all order statuses (Admin)",
        "operationId": "GetAllOrderStatuses",
        "responses": {
          "200": {
            "description": "Successfully retrieved all order statuses",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "204": {
            "description": "No order statuses found",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to find all order statuses",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/orders/{shop_order_id}/items": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to get all order items of a specific order",
        "tags": ["Admin Orders"],
        "summary": "Get all order items (Admin)",
        "operationId": "GetAllOrderItemsAdmin",
        "parameters": [
          {
            "type": "integer",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found order items",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "204": {
            "description": "No order items found",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to find order items",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/carts/place-order": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user save an order",
        "tags": ["User Orders"],
        "summary": "Save Order (User)",
        "operationId": "SaveOrder",
        "parameters": [
          {
            "type": "string",
            "description": "Address ID",
            "name": "address_id",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully order placed",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "204": {
            "description": "Cart is empty",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "409": {
            "description": "Can't place order out of stock product on cart",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to save order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API to get order for user user orders",
        "tags": ["User Orders"],
        "summary": "Get user orders (User)",
        "operationId": "GetUserOrder",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count Of Order",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all user orders",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "204": {
            "description": "No shop orders for user",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to retrieve all user orders",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/orders/return": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to request a return for delivered order",
        "tags": ["User Orders"],
        "summary": "Return request (User)",
        "operationId": "SubmitReturnRequest",
        "parameters": [
          {
            "description": "Input Fields",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.Return"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully return request submitted for order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/orders/{shop_order_id}/cancel": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Api for user to cancel a order",
        "tags": ["User Orders"],
        "summary": "Cancel order (User)",
        "operationId": "CancelOrder",
        "parameters": [
          {
            "type": "integer",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully order cancelled",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "Invalid inputs",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to cancel order",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/orders/{shop_order_id}/items": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to get all order items of a specific order",
        "tags": ["User Orders"],
        "summary": "Get all order items (User)",
        "operationId": "GetAllOrderItemsUser",
        "parameters": [
          {
            "type": "integer",
            "description": "Shop Order ID",
            "name": "shop_order_id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count Of Order",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found order items",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to find order items",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "request.Return": {
      "type": "object",
      "required": ["return_reason", "shop_order_id"],
      "properties": {
        "return_reason": {
          "type": "string",
          "maxLength": 150,
          "minLength": 6
        },
        "shop_order_id": {
          "type": "integer"
        }
      }
    },
...
    "request.UpdateOrder": {
      "type": "object",
      "required": ["shop_order_id"],
      "properties": {
        "order_status_id": {
          "type": "integer"
        },
        "shop_order_id": {
          "type": "integer"
        }
      }
    },
    "request.UpdateOrderReturn": {
      "type": "object",
      "required": ["admin_comment", "order_return_id", "order_status_id"],
      "properties": {
        "admin_comment": {
          "type": "string",
          "maxLength": 150,
          "minLength": 6
        },
        "order_return_id": {
          "type": "integer"
        },
        "order_status_id": {
          "type": "integer"
        },
        "return_date": {
          "type": "string"
        }
      }
    },
...
    "response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "error": {},
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}